refactor
Migrated a large chunk of the editor specific code into the NGO editor assembly.
