+++ com.unity.netcode.gameobjects/Editor/NetworkManagerHelper.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9f88c8d8116709e4dbc5909fc0415fac
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
